requirements.txt
numpy==1.19.5
python-gdcm         // The library must be installed than pydicom
pydicom==2.0.0
tensorflow-gpu==2.2.0
scikit-learn==0.22.1
opencv-python==4.5.1.48
Django==3.1.1
nibabel==3.2.1
scipy=1.4.1
pandas==1.0.1
